Emails: Move PriceBadge component to new folder
