+++ client/my-sites/email/email-providers-comparison/price-badge/style.scss
@@ -1,0 +1,8 @@
+@import '@automattic/calypso-color-schemes';
+
+.price-badge {
+	.promo-card__price .price__discount {
+		text-decoration: line-through;
+		color: var( --studio-gray-60 );
+	}
+}
