+++ client/my-sites/email/email-providers-stacked-comparison/provider-cards/google-workspace-card.tsx
@@ -22,8 +22,8 @@ import {
 } from 'calypso/lib/gsuite/new-users';
 import { formatPrice } from 'calypso/lib/gsuite/utils/format-price';
 import withCartKey from 'calypso/my-sites/checkout/with-cart-key';
+import PriceBadge from 'calypso/my-sites/email/email-providers-comparison/price-badge';
 import EmailProvidersStackedCard from 'calypso/my-sites/email/email-providers-stacked-comparison/email-provider-stacked-card';
-import PriceBadge from 'calypso/my-sites/email/email-providers-stacked-comparison/provider-cards/price-badge';
 import PriceWithInterval from 'calypso/my-sites/email/email-providers-stacked-comparison/provider-cards/price-with-interval';
 import {
 	EmailProvidersStackedCardProps,
