+++ client/my-sites/email/email-providers-stacked-comparison/provider-cards/professional-email-card.scss
@@ -12,11 +12,6 @@
 	}
 }
 
-.professional-email-card .provider-cards__price-badge .promo-card__price .price__discount {
-		text-decoration: line-through;
-		color: var( --studio-gray-60 );
-}
-
 .professional-email-card__discount.badge {
 	margin-bottom: 12px;
 }
