+++ client/my-sites/email/email-providers-stacked-comparison/provider-cards/professional-email-card.tsx
@@ -24,8 +24,8 @@ import {
 	validateMailboxes as validateTitanMailboxes,
 } from 'calypso/lib/titan/new-mailbox';
 import withCartKey from 'calypso/my-sites/checkout/with-cart-key';
+import PriceBadge from 'calypso/my-sites/email/email-providers-comparison/price-badge';
 import EmailProvidersStackedCard from 'calypso/my-sites/email/email-providers-stacked-comparison/email-provider-stacked-card';
-import PriceBadge from 'calypso/my-sites/email/email-providers-stacked-comparison/provider-cards/price-badge';
 import PriceWithInterval from 'calypso/my-sites/email/email-providers-stacked-comparison/provider-cards/price-with-interval';
 import {
 	TITAN_PASSWORD_RESET_FIELD,
